Created Init class for initializing engine features. Extracted more engine code.

+++ v1/Engine/include/WindowManager.h
@@ -13,7 +13,7 @@ namespace Engine
 		SDL_Window* window;																			// Declare SDL_Window*
 
 	public:																							// Public member functions declared below.
-		WindowManager(const char* windowTitle, const int pixelWidth, const int pixelHeight);		// Declare WindowManager constructor.
+		WindowManager(const char* windowTitle, const uint16_t pixelWidth, const uint16_t pixelHeight);		// Declare WindowManager constructor.
 		SDL_Window* getWindow() const;																// Declare getWindow function.
 		void free();																				// Declare free function.
 	};
